// src/api/model/createTaskDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vector
 * The Vector API
 * OpenAPI spec version: 1.0
 */
import type { CreateTaskDtoNotes } from './createTaskDtoNotes';

export interface CreateTaskDto {
  workspaceId: number;
  sourceId: number;
  title: string;
  description?: string;
  flagged?: boolean;
  deadlineType: string;
  issuedAt?: string;
  dueDate: string;
  notes?: CreateTaskDtoNotes;
}